Split lib/types/game.ts into barrel-exported domain files
Break the oversized src/lib/types/game.ts (~471 lines, over the 400-line cap)
into per-domain files under src/lib/types/game/ (game, lobby, mode, player,
mode-services) re-exported from an index barrel, so the @/lib/types import path
is unchanged and no consumer breaks. Update docs/GETTING_STARTED.md references
to the new game/ directory. Behavior-preserving refactor.
