+++ src/lib/types/game/mode-services.ts
@@ -1,166 +1,10 @@
-import type {
-  AvalonLobbyConfig,
-  AvalonModeConfig,
-} from "@/lib/game/modes/avalon/lobby-config";
-import type {
-  ClocktowerLobbyConfig,
-  ClocktowerModeConfig,
-} from "@/lib/game/modes/clocktower/lobby-config";
-import type {
-  CodenamesLobbyConfig,
-  CodenamesModeConfig,
-} from "@/lib/game/modes/codenames/lobby-config";
-import type {
-  SecretVillainLobbyConfig,
-  SecretVillainModeConfig,
-} from "@/lib/game/modes/secret-villain/lobby-config";
-import type { SecretVillainTimerConfig } from "@/lib/game/modes/secret-villain/timer-config";
-import type {
-  WerewolfLobbyConfig,
-  WerewolfModeConfig,
-} from "@/lib/game/modes/werewolf/lobby-config";
-import type { WerewolfTimerConfig } from "@/lib/game/modes/werewolf/timer-config";
-
-import type { GameStatusState } from "./game-status";
-import type { LobbyPlayer, TimerConfig } from "./lobby";
-import type { ModeConfig } from "./mode-config";
-import type { RoleDefinition, Team } from "./role";
-import type { RoleBucket } from "./role-bucket";
-
-export {
-  type FinishedGameStatus,
-  GameStatus,
-  type GameStatusState,
-  type GameWinner,
-  type PlayingGameStatus,
-  type StartingGameStatus,
-} from "./game-status";
-export type {
-  DeviceLobbyPlayer,
-  LobbyPlayer,
-  NoDeviceLobbyPlayer,
-  TimerConfig,
-} from "./lobby";
-export { DEFAULT_TIMER_CONFIG } from "./lobby";
-export { type RoleDefinition, Team } from "./role";
-
-// --- Game Modes ---
-
-export enum GameMode {
-  Avalon = "avalon",
-  Clocktower = "clocktower",
-  Codenames = "codenames",
-  SecretVillain = "secret-villain",
-  Werewolf = "werewolf",
-}
-
-// --- Role Config Modes ---
-
-export enum RoleConfigMode {
-  Default = "Default",
-  Custom = "Custom",
-  Advanced = "Advanced",
-}
-
-// --- Show Roles In Play Options ---
-
-export enum ShowRolesInPlay {
-  None = "None",
-  ConfiguredOnly = "ConfiguredOnly",
-  AssignedRolesOnly = "AssignedRolesOnly",
-  RoleAndCount = "RoleAndCount",
-}
-
-export interface PlayerRoleAssignment {
-  playerId: string;
-  roleDefinitionId: string;
-}
-
-export type VisibilityReason = "wake-partner" | "aware-of";
-
-export interface VisiblePlayer {
-  playerId: string;
-  reason: VisibilityReason;
-  /** When set, the player's role is known (not just their identity). */
-  roleId?: string;
-}
-
-export type GamePlayer = LobbyPlayer & { visiblePlayers: VisiblePlayer[] };
-
-// --- Game (exists only after the game has been started) ---
-
-/** Shared game fields. Game-mode-specific variants extend this. */
-export interface BaseGame {
-  id: string;
-  lobbyId: string;
-  status: GameStatusState;
-  players: GamePlayer[];
-  roleAssignments: PlayerRoleAssignment[];
-  configuredRoleBuckets: RoleBucket[];
-  showRolesInPlay: ShowRolesInPlay;
-  ownerPlayerId?: string;
-  /** Executioner: the player ID the Executioner must get eliminated at trial. */
-  executionerTargetId?: string;
-  /** Lobby seating order: player IDs in position order, used to set president rotation in Secret Villain. */
-  playerOrder?: string[];
-}
-
-export interface WerewolfGame extends BaseGame {
-  gameMode: GameMode.Werewolf;
-  timerConfig: WerewolfTimerConfig;
-  modeConfig: WerewolfModeConfig;
-  /** Role IDs that were drawn but not assigned to players (hidden from everyone except the Narrator). */
-  hiddenRoleIds?: string[];
-}
-
-export interface SecretVillainGame extends BaseGame {
-  gameMode: GameMode.SecretVillain;
-  timerConfig: SecretVillainTimerConfig;
-  modeConfig: SecretVillainModeConfig;
-}
-
-export interface AvalonGame extends BaseGame {
-  gameMode: GameMode.Avalon;
-  timerConfig: TimerConfig;
-  modeConfig: AvalonModeConfig;
-}
-
-export interface ClocktowerGame extends BaseGame {
-  gameMode: GameMode.Clocktower;
-  timerConfig: TimerConfig;
-  modeConfig: ClocktowerModeConfig;
-  /**
-   * The Townsfolk role ID shown to the Drunk player as their fake token.
-   * Assigned at game creation; only present when the Drunk is in play.
-   */
-  drunkFakeRoleId?: string;
-}
-
-export interface CodenamesGame extends BaseGame {
-  gameMode: GameMode.Codenames;
-  timerConfig: TimerConfig;
-  modeConfig: CodenamesModeConfig;
-}
-
-/**
- * Discriminated union of all game-mode-specific game objects.
- * Narrow on `gameMode` to access typed `timerConfig` and `modeConfig`.
- */
-export type Game =
-  | WerewolfGame
-  | SecretVillainGame
-  | AvalonGame
-  | ClocktowerGame
-  | CodenamesGame;
-
-/**
- * A game-mode-defined action that can be applied to a game. Actions are
- * validated before application; isValid must return true for apply to be called.
- */
-export interface GameAction {
-  isValid(game: Game, callerId: string, payload: unknown): boolean;
-  apply(game: Game, payload: unknown, callerId: string): void;
-}
+import type { TimerConfig } from "../lobby";
+import type { ModeConfig } from "../mode-config";
+import type { RoleDefinition, Team } from "../role";
+import type { RoleBucket } from "../role-bucket";
+import type { Game, GameAction } from "./game";
+import type { BaseLobbyConfig, LobbyConfig } from "./lobby";
+import type { PlayerRoleAssignment } from "./player";
 
 /**
  * Game-mode-specific service methods for state extraction and initialization.
@@ -298,38 +142,4 @@ export interface GameModeConfig {
   buildDefaultLobbyConfig(base: BaseLobbyConfig): LobbyConfig;
   readonly actions: Record<string, GameAction>;
   readonly services: GameModeServices;
-}
-
-// --- Lobby (top-level entity; game is absent until started) ---
-
-/** Shared lobby config fields. Game-mode-specific variants extend this. */
-export interface BaseLobbyConfig {
-  roleConfigMode: RoleConfigMode;
-  roleBuckets: RoleBucket[];
-  showConfigToPlayers: boolean;
-  showRolesInPlay: ShowRolesInPlay;
-}
-
-/**
- * Discriminated union of all game-mode-specific lobby configurations.
- * Narrow on `gameMode` to access typed `timerConfig` and `modeConfig`.
- */
-export type LobbyConfig =
-  | WerewolfLobbyConfig
-  | SecretVillainLobbyConfig
-  | AvalonLobbyConfig
-  | ClocktowerLobbyConfig
-  | CodenamesLobbyConfig;
-
-export interface Lobby {
-  id: string;
-  ownerSessionId: string;
-  players: LobbyPlayer[];
-  /** Ordered list of player IDs defining seating positions (index 0 = first seat). */
-  playerOrder: string[];
-  config: LobbyConfig;
-  gameId?: string;
-  readyPlayerIds: string[];
-  /** Unix ms timestamp set server-side when all players (including owner) ready up. Cleared on unready, player leave, or game start. */
-  countdownStartedAt?: number;
 }
